Wed, Jan 3, 2024, 11:52 PM +08:00

+++ readability/readability.c
@@ -19,9 +19,9 @@ int main(void)
     int sentences = count_sentences(text);
 
     // Compute the Coleman-Liau index
-    double L = ((float) letters / (float) words) * 100;
-    double S = ((float) sentences / (float) words) * 100;
-    double index = (0.0588 * L) - (0.296 * S) - 15.8;
+    float L = ((float) letters / (float) words) * 100;
+    float S = ((float) sentences / (float) words) * 100;
+    float index = (0.0588 * L) - (0.296 * S) - 15.8;
 
     // Print the grade level
     if ((int) round(index) >= 16)
